requirements.txt
PyPDF2
pillow
pandas
pdf2image
google-generativeai
python-dotenv
absl-py
